Create basic slim bundler api
- Removes custom concat logic
- Single template for the slim loader with conditional pieces
- Adds `graph/slim` which takes care of mutating the bundle nodes
- Adds a separate stream to call `graph/slim` with the right arguments

Closes #666
