+++ lib/node/make_slim_config_node.js
@@ -2,10 +2,8 @@ var path = require("path");
 var prettier = require("prettier");
 var getBundleFileName = require("../bundle/filename");
 
-module.exports = function(data) {
-	var loader = data.loader;
-	var bundles = data.bundles;
-	var configuration = data.configuration;
+module.exports = function(options) {
+	var bundles = options.bundles;
 
 	// maps slim module ids (numbers) to slim bundle ids
 	var slimBundlesConfig = {};
@@ -16,11 +14,21 @@ module.exports = function(data) {
 	// maps slim bundle ids to the bundle file address
 	var slimPathsConfig = {};
 
+	/**
+	 * Retuns bundles path relative to the loader baseUrl
+	 *
+	 * The bundles path at this point is an absolute file url, the relative
+	 * version is needed in the written config so the loader can use script
+	 * tags to load the bundles during runtime.
+	 *
+	 * @param {Object} bundle - The bundle object
+	 * @return {string} the relative path
+	 */
 	function getRelativeBundlePath(bundle) {
-		var baseUrl = loader.baseURL.replace("file:", "");
+		var baseUrl = options.baseUrl.replace("file:", "");
 
 		return path.join(
-			path.relative(baseUrl, configuration.bundlesPath),
+			path.relative(baseUrl, options.bundlesPath),
 			getBundleFileName(bundle)
 		);
 	}
@@ -35,8 +43,8 @@ module.exports = function(data) {
 		});
 	});
 
-	loader.bundle.forEach(function(bundleName) {
-		var node = data.graph[bundleName];
+	options.progressiveBundles.forEach(function(bundleName) {
+		var node = options.graph[bundleName];
 		slimMapConfig[bundleName] = node.load.uniqueId;
 	});
