+++ lib/node/make_slim_shim_node.js
@@ -1,5 +1,7 @@
 var prettier = require("prettier");
-var makeShimCode = require("./slim/make_shim_code");
+var endsWith = require("lodash/endsWith");
+var getNodeSource = require("./source").node;
+var makeShimTemplate = require("./slim/make_shim_template");
 
 module.exports = function(options) {
 	// Ignore some nodes not needed for the slim loader
@@ -7,11 +9,21 @@ module.exports = function(options) {
 	// nodes[1] === @configMain
 	var modules = options.nodes.slice(2);
 
+	var template = makeShimTemplate({
+		progressive: options.progressive
+	});
+
+	var args = modules
+		.map(function(node) {
+			var code = getNodeSource(node).code.toString();
+			return endsWith(code, ";") ? code.substring(0, code.length - 1) : code;
+		})
+		.join(",");
+
 	var code = prettier.format(
-		makeShimCode({
-			modules: modules,
-			mainModuleId: options.mainModuleId,
-			multipleBundles: options.multipleBundles
+		template({
+			args: args,
+			mainModuleId: options.mainModuleId
 		}),
 		{ useTabs: true }
 	);
